cargo: Set `rust-version` to keep in sync with the CI MSRV
Provoke users of the template to set the `rust-version` in `Cargo.toml`
for a multitude of reasons (crates.io shows it to users, and `cargo`
can use it in its MSRV-aware resolver), including the `+nightly
-Zminimal-versions generate-lockfile` step to generate a compatible
`version = 3` `Cargo.lock` for MSRV 1.74 to consume, rather than
`version = 4` which it doesn't understand yet.

This also supersedes the old "hack" to install an older `nightly`
version that still generated `version = 3` lockfiles.