bindgen: Generate Rustified enums with repetitive prefixes stripped
The common prefix stripping of enums was only implemented for
`FFX_RESOURCE_STATE_`; implement it generically for everything with some
workarounds to deal with odd typename suffixes and broken `CamelCase`
suffixes.

Note one very important thing: Rustified enums are _awesome_ when
passing arguments **into FFI** but are UB when returned; there's no
guarantee an external function doesn't return a new number (after
updating the code, or perhaps it was forgotten in the `enum`) and this
is UB in Rust.  For those types, such as `ErrorCodes` a newtype enum
has been generated.  In this special case the underlying implementation
doesn't return that but a `type ErrorCode = i32`, which we instead map
to an opaque wrapper with appropriate `Display`/`Debug`/`Error`.