Delete unusable `fidelityfx` high-level wrapper crate
It is unclear what was attempted when the high-level bindings were
written, but it's clear that they're unfinished and defy the safe
definitions of Rust and lifetimes.  For one because this is incredibly
hard to define when the calling types don't clearly define ownership
(a borrow to a clonable Dx12 `ComPtr` is not usable, a borrow of an
untracked `ash::vk::CommandBuffer` handle for example is meaningless).
Secondly because the underlying FidelityFX SDK has its fair share
of problems, like keeping pointers "alive" by copying the pointee
to static globals and returning pointers to that which is completely
not reentrant.

All in all, we best delete the high-level crate rather than
painstakingly fixing or rewriting it, providing only the low-level
generated `sys` bindings that require all callers to confirm and uphold
whichever safety (UB-preventing) requirements in their own way.