Replace `FidelityFX-SDK` copy with an actual submodule
The SDK is massive, and is going to be unnecessary beyond bindings
generation after we learned (and still have to change in a future
PR) that we are not supposed to **manually build** the CPP files and
statically link into our app, but rather dynamically link against (or in
our case open at runtime via `libloading`) the DLLs from the "package"
provided on GPUOpen (https://gpuopen.com/amd-fidelityfx-sdk-1/ and soon
https://gpuopen.com/amd-fidelityfx-sdk/ for V2 with FSR4).

Note also that the DLL structure explains how to incorporate Vulkan and
Dx12 concurrently; the `shader_permutations` for `vk` and `dx12` have
the exact same global static symbol names making it impossible to embed
both sources at the same time with successful static linking (while our
`build.rs` has thus far shown that we can compile and link the rest of
the sources concurrently), so upstream provides two independent DLLs
that should be loaded separately from each other.

The submodule is currently checked out to the [`release-FSR3-3.0.3`]
tag, matching the original source we had as demonstrated by not having
any bindings changes.

[`release-FSR3-3.0.3`]: https://github.com/GPUOpen-LibrariesAndSDKs/FidelityFX-SDK/tree/release-FSR3-3.0.3