Filter unused external header noise from `bindings.rs`
Unlike the Vulkan and Dx12 bindings, the root "bindings" include
everything **through FSR3** with no filtering whatsoever (because a
file-based filter might be hairy).  In a first attempt to hide unused
external header noise, put some `allowlist` rules in place to only
accept FSR types.

As soon as we actually generate a separate bindings file for FSR3
and the Agility SDK "core", we might be able to simplify these filters
with a simple file-filter again instead.