sys: Move all current bindings to an `sdk` crate module (#45)
We realized that the recent upgrade to FFX 1.1.4 includes a completely
new `ffx-api` that is supposed to be used over the older SDK bindings,
especially after removing in-tree compilation of the SDK too. Move the
bindings to an `sdk` module to start the separation, pending full
removal if it all turns out to no longer be useful at all.