Replace helpless `Default` for header structs with proper `type_` initialization
It's quite common to forget to initialize `self.header.type_`; in "safe"
Rust these sort of things should just be taken care of for you.

Mimic `ash`'s `TaggedStructure` trait which also helps take care of
this, together with a safe `Header` getter rather than pointer-casting
(even though this could be made a requirement of the `unsafe trait`).