Implement missing `Result` conversion for `ReturnCode` wrapper
Despite implementing `std::error::Error` for all of `ReturnCode`, the
`OK` variant should not be treated as an `Err` case; add a helper `fn
result()` getter to extract these as `Ok` and `Err(self)` respectively.

It would be impossible/inconvenient to create an error-only `ReturnCode`
variant without `OK`, because it's simply a newtype wrapper around some
bindgen constants.