Regenerate bindings against `v2.1.0`
This upgrades us away from `1.1.4` to the new [`2.1.0` SDK](
https://github.com/GPUOpen-LibrariesAndSDKs/FidelityFX-SDK/releases/tag/v2.1.0),
with quite some changes required to deal with `DESC_TYPE` constants
being expressions rather than constants and the like, but in the end the
public API remains almost identical.

The only new binding for FSR 4 is `ffxCreateContextDescUpscaleVersion`
since `2.1.0`, "required" to "acknowledge" (I guess?) to the
implementation that we really want to opt into FSR 4 although it was/is
already working with an explicit `version_override`; the duplicate
system using an (as of writing) hardcoded header constant version
`4.1.1` is somewhat confusing.

On our end thanks to split headers we have now moved the `dx12` bindings
for `framegeneration` from the "root" DX12 backend bindings into a new
Rust `framegeneration::dx12::` module.

Notably Vulkan supports has entirely been removed from the SDK, and
hence removed from this crate as well beyond a few straddlers left in
the upstream headers for "backwards compatibility" (?).